requirements.txt
google-generativeai
pyttsx3
SpeechRecognition
pyaudio
Pillow
spotipy
requests
streamlit
streamlit-webrtc
streamlit-autorefresh
mediapipe
opencv-python
